Move noInteractionButton to styles
