+++ userinterface/Styles/NoInteractionButtonView.axaml
