+++ userinterface/Views/Controls/EditableExpanderView.axaml.cs
@@ -12,6 +12,7 @@ using System.ComponentModel;
 using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
+using Styles;
 
 namespace userinterface.Controls
 {
